Create "LoginServlet" who is response for login an User.Check whether user is exist and  place login ore interrupted login.

+++ JSPAndServletsBank/src/main/java/com/clouway/core/AccountBankDAO.java
@@ -5,8 +5,8 @@ package com.clouway.core;
  */
 public interface AccountBankDAO {
 
-  void deposit(int amount, int userID);
+  void deposit(float amount, int userID);
 
-  int withdrawing(int withdrawingAmount, int userID);
+  int withdrawing(float withdrawingAmount, int userID);
 
 }
